test(vapor): move valueless attr regression tests to own module
Keeps tests.rs from growing past its over-limit baseline, which the
source-file-lengths guard rejects. The two regression tests move to a new
tests_valueless_attr module with matching snapshot filenames.
