+++ crates/vize_atelier_vapor/src/tests.rs
@@ -901,42 +901,3 @@ fn test_compile_v_memo_with_dependencies_reports_diagnostic() {
     assert!(!result.code.contains("_withDirectives"), "{}", result.code);
     assert!(!result.code.contains("_memo"), "{}", result.code);
 }
-
-// Regression tests: a valueless static attribute on a component is an empty
-// string prop (matching vdom), not `undefined`, which would drop the
-// attribute from the rendered element.
-#[test]
-fn test_compile_component_valueless_static_attr_is_empty_string_prop() {
-    let allocator = Bump::new();
-    let result = compile_vapor(&allocator, r#"<MyComp data-probe />"#, Default::default());
-
-    assert!(
-        result.error_messages.is_empty(),
-        "Expected no errors: {:?}",
-        result.error_messages
-    );
-
-    let code = normalize_code(&result.code);
-    assert_parses_as_module(&code);
-    insta::assert_snapshot!(code.as_str());
-}
-
-#[test]
-fn test_compile_dynamic_component_valueless_static_attr_is_empty_string_prop() {
-    let allocator = Bump::new();
-    let result = compile_vapor(
-        &allocator,
-        r#"<component :is="tag" data-probe>hi</component>"#,
-        Default::default(),
-    );
-
-    assert!(
-        result.error_messages.is_empty(),
-        "Expected no errors: {:?}",
-        result.error_messages
-    );
-
-    let code = normalize_code(&result.code);
-    assert_parses_as_module(&code);
-    insta::assert_snapshot!(code.as_str());
-}
